include/fi_shm: Rename to use ofi file name prefix
This is an internal header, use the ofi prefix instead of
fi, which is to be reserved for external APIs only.
