+++ prov/shm/src/smr.h
@@ -53,7 +53,7 @@
 
 #include <fi.h>
 #include <fi_enosys.h>
-#include <fi_shm.h>
+#include <ofi_shm.h>
 #include <ofi_rbuf.h>
 #include <fi_list.h>
 #include <ofi_signal.h>
